Extract Geolocation out of CommonUtils
First step of retiring CommonUtils as a kitchen-sink class.

`format_insdc_latlon`, `convert_latlon_insdc2iso`, `country_at` and
their internals (`point_in_polygon?`, `point_in_ring?`, plus the
lazily-loaded Natural Earth / INSDC-ISO_A3 data) are all related and
pure functions — move them into a new `Geolocation` module that holds
its own data caches as module-level instance variables. No more
`CommonUtils.new.format_insdc_latlon(...)` dance, just
`Geolocation.format_insdc_latlon(...)`.

Callers updated:
- lib/validator/biosample_validator.rb: 3 validator methods
  (invalid_lat_lon_format, invalid_lat_lon, latlon_versus_country)
  drop the `common = CommonUtils.new` temp and call
  `Geolocation.*` directly.
- test/lib/validator/common/common_utils_test.rb: 3 test cases
  re-target `Geolocation` too.
- add `require .../common/geolocation.rb` where needed.

lib/validator/common/common_utils.rb: drop the ~140-line geolocation
block (including the NE_COUNTRIES_PATH / INSDC_TO_ISO_A3_PATH constants
and the two lazy loaders). No other behavior change.

Remaining `CommonUtils.new` call sites use other CommonUtils methods
(parse_coll_dump, exist_pubmed_id?, http_get_response) and will come
off as their domains get extracted in follow-up PRs.

Verified: 324 runs / 2600 assertions / 0 failures.
